requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
pydantic>=2.6,<3
openai>=1.30,<2
python-dotenv>=1.0,<2
yfinance>=0.2.30
